Updated connectal code and macros for selecting different platforms.
By default, the code assums a spike-like platform with a boot rom in the
bottom part of the address space and ram starting at address 0x80000000.
This platform uses shared memory in connectal for ram and rom.

If CONFIG_PLATFORM_MCU is defined, the code targets a platform for a
microcontroller with an SRAM at the begining of the address space.
This platform uses an external memory interface to load that SRAM.